requirements.txt
dill~=0.3.8
uvicorn==0.29.0
fastapi==0.110.0
openpyxl==3.1.2
Unidecode==1.3.3
workalendar==17.0.0
xlrd==2.0.1
holiday
